added user_state_clent dir
